% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_backfillz.R
\name{as_backfillz}
\alias{as_backfillz}
\title{Convert stanFit object to a Backfillz object}
\usage{
as_backfillz(object, verbose = TRUE)
}
\arguments{
\item{object}{Stanfit or dataframe.}

\item{verbose}{Set to TRUE to see function messages
Dataframe must contain columns 'parameters', 'iterations' and 'chains'.}
}
\description{
Convert stanFit object to a Backfillz object
}
