% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfillz-Class.R
\docType{class}
\name{Backfillz-class}
\alias{Backfillz-class}
\alias{Backfillz}
\title{An S4 class to represent a the backfillz plot object.}
\description{
An S4 class to represent a the backfillz plot object.
}
\section{Slots}{

\describe{
\item{\code{theme_name}}{Name of current theme}

\item{\code{theme_text_family}}{Text family theme option for plots}

\item{\code{theme_text_font}}{Text font theme option for plots}

\item{\code{theme_text_font_colour}}{Text font colour theme option for plots}

\item{\code{theme_text_cex_title}}{Text size title for plots}

\item{\code{theme_text_cex_main}}{Text size main for plots}

\item{\code{theme_text_cex_axis}}{Text size axis for plots}

\item{\code{theme_text_col_title}}{Text colour title for plots}

\item{\code{theme_text_col_main}}{Text colour main for plots}

\item{\code{theme_text_col_axis}}{Text colour axis for plots}

\item{\code{theme_bg_colour}}{Background colour theme option for plots}

\item{\code{theme_mg_colour}}{Middleground colour theme option for plots}

\item{\code{theme_fg_colour}}{Foregound colour option for plots}

\item{\code{theme_alpha}}{Transparency option for plots}

\item{\code{theme_palette}}{List of hexdecimal colours for use in plots}

\item{\code{mcmc_samples}}{MCMC samples from a bayesian model fit}

\item{\code{mcmc_model}}{Text definition of MCMC model}

\item{\code{plot_history}}{Data frame with overall plot history}

\item{\code{plot_store}}{List containing stored plots}

\item{\code{df_slice_histogram}}{Slot containing data from plot_slice_histrogram function}

\item{\code{df_trace_dial}}{Slot containing data from plot_slice_dial function}

\item{\code{df_spiral_stream}}{Slow containing data from plot_spiral_stream function}
}}

