% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_stanarray.R
\name{df_to_stanarray}
\alias{df_to_stanarray}
\title{Converts a data frame to the output structure of extract}
\usage{
df_to_stanarray(df)
}
\arguments{
\item{df}{data frame with the column names
'parameters', 'iterations' and 'chains'}
}
\value{
An array from df with the
 dimensions [iteration, chain, parameters]
}
\description{
Converts a data frame to the output structure of extract
}
\examples{
df <- data.frame(
              parameters = c('mu', 'mu'),
              iterations = c(0.345, 0.263),
              chains = c(1, 2),
              stringsAsFactors = FALSE
              )
df_to_stanarray(df)
}
