% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfillz-Class.R
\name{plot}
\alias{plot}
\title{Plot method for the backfillz method}
\usage{
\S4method{plot}{Backfillz,missing}(x, y, ...)
}
\arguments{
\item{x}{Backfillz object}

\item{PlotType}{Type of plot to create. Either 'slice_histogram', 'trace_dial' or 'spiral_stream'. Defaults to 'slice_histogram'.}
}
\description{
Plot method for the backfillz method
}
