% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spiral_stream.R
\name{plot_spiral_stream}
\alias{plot_spiral_stream}
\title{Creates a spiral steamgrapgh of MCMC chains}
\usage{
plot_spiral_stream(
  object = NULL,
  start_sample = NULL,
  steps = NULL,
  parameters = NULL,
  laps = 4,
  save_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Backfillz or Stanfit object. Stanfit objects
are converted to Backfillz object using the as_backfillz function.}

\item{start_sample}{Chain position to start the spirals.}

\item{steps}{Vector of step increments for each spiral
within chain. Defaults to 25 percent of sample for 4 steps.}

\item{parameters}{Vector of parameters to plot. Defaults to
two parameters.}

\item{laps}{Number of laps for the spitals. Defaults to 4.}

\item{save_plot}{Set to TRUE to save plots in the Backfillz object.}

\item{verbose}{Set to TRUE to see function messages}
}
\description{
Creates a spiral steamgrapgh of MCMC chains
}
