% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace_dial.R
\name{plot_trace_dial}
\alias{plot_trace_dial}
\title{Creates a trace dial plot}
\usage{
plot_trace_dial(
  object,
  parameters = NULL,
  n_bins = 40,
  n_burnin = 10,
  save_plot = FALSE
)
}
\arguments{
\item{object}{Backfillz or Stanfit object. Stanfit objects are converted to Backfillz object using the as_backfillz function.}

\item{parameters}{Vector of strings giving the parameters to plot (e.g., c('mu', 'sd')).}

\item{n_bins}{Number of bins displayed in histograms at end of dial.}

\item{n_burnin}{Number of burnin samples.}

\item{save_plot}{Set to TRUE to save plots in the Backfillz object.}
}
\description{
Creates a trace dial plot
}
