% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_stanfit}
\alias{sample_stanfit}
\title{Example stanfit object}
\format{
A stanfit object with 18 parameters,
 3 chains, 10000 iterations and 500 warmup samples
}
\usage{
sample_stanfit
}
\description{
An example stanfit object following the example fit
 of a heirarchical schools model from the
 pystan documentation.
 'https://pystan.readthedocs.io/en/latest/'.
}
\examples{
# show model used to create sample
sample_stanfit@stanmodel
# print model summary
sample_stanfit
}
\keyword{datasets}
