% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme.R
\name{set_theme}
\alias{set_theme}
\title{Set graphical plot slots of a backfillz object to a theme}
\usage{
set_theme(backfillz_object, theme = "default")
}
\arguments{
\item{backfillz_object}{Backfillz object}

\item{theme}{Character specifying theme
(e.g., 'default', 'solarized_dark')}
}
\description{
Set graphical plot slots of a backfillz object to a theme
}
