% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_backfillz.R
\name{as_backfillz}
\alias{as_backfillz}
\title{Convert stanFit object to a Backfillz object}
\usage{
as_backfillz(stanfit_object)
}
\arguments{
\item{stanfit_object}{Backfillz object}
}
\description{
Convert stanFit object to a Backfillz object
}
