% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_stanfit}
\alias{sample_stanfit}
\title{Example stanfit object}
\format{
A stanfit object with 18 parameters, 3 chains, 10000 iterations and 500 warmup samples
}
\usage{
sample_stanfit
}
\description{
An example stanfit object.
}
\keyword{datasets}
